Move ManualControll class to rendering.py
